requirements.txt
pytorch==1.0.0
torchvision=0.2.1
keras